+++ src/RestApi/resources.ts
@@ -1,0 +1,5 @@
+export enum Resources {
+	Device = 'device',
+	Location = 'location',
+	OrganizationTag = 'organization-tag',
+}